rag_pipeline/requirements.txt
# Core dependencies for Government Schemes RAG Pipeline

# Embeddings
sentence-transformers>=2.2.0

# Vector database
pinecone>=3.0.0

# LLM - Groq (free, fast inference)
groq>=0.4.0

# Environment and utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
